move heightData to fix build issue
